test: use new endpoint to post submission skipping enqueuing evaluation and run orchestrator job.
